refactor(focus): extract handlers; route becomes a 48-line shim
Moves prompt selection, parallel sub-component fan-out, the
calibration-suggestions side-channel, fallback shape, and entitlement
re-throw out of the route file and into src/lib/focus/handlers.ts.

Route /api/focus now does only: parse body, call withGuardedRoute,
delegate to generateFocus. From 310 LOC -> 48 LOC.
